api/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]
langchain
openai
langchain-openai
python-dotenv